split time-consuming UTs for CPU CI time usage
