+++ test/test_cpu/core/test_autoround.py
@@ -438,300 +438,3 @@ class TestAutoRound:
         text = "There is a girl who likes adventure,"
         inputs = tokenizer(text, return_tensors="pt").to(model.device)
         res = tokenizer.decode(model.generate(**inputs, max_new_tokens=1)[0])
-
-    def test_fallback_layers_regex_awq(self, tiny_opt_model_path, dataloader):
-        model_name = tiny_opt_model_path
-        bits, group_size, sym = 4, 128, True
-        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
-        tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
-        layer_config = {
-            r"model\.decoder\.layers\.(?:[0-9]|1[0-1])\.self_attn\.q_proj": {"bits": 16},
-            "model.decoder.layers.1.self_attn.k_proj": {"bits": 16},
-        }
-        autoround = AutoRound(
-            model,
-            tokenizer=tokenizer,
-            bits=bits,
-            group_size=group_size,
-            sym=sym,
-            iters=2,
-            seqlen=2,
-            dataset=dataloader,
-            layer_config=layer_config,
-        )
-        autoround.quantize()
-        quantized_model_path = self.save_folder
-
-        _, quantized_model_path = autoround.save_quantized(
-            output_dir=quantized_model_path, format="auto_awq", inplace=True, return_folders=True
-        )
-        quantization_config = AutoRoundConfig()
-
-        model = AutoModelForCausalLM.from_pretrained(
-            quantized_model_path, device_map="auto", quantization_config=quantization_config
-        )
-        tokenizer = AutoTokenizer.from_pretrained(quantized_model_path)
-        text = "There is a girl who likes adventure,"
-        inputs = tokenizer(text, return_tensors="pt").to(model.device)
-        res = tokenizer.decode(model.generate(**inputs, max_new_tokens=5)[0])
-        print(res)
-
-    def test_fallback_layers_regex_gptq(self, tiny_opt_model_path, dataloader):
-        model_name = tiny_opt_model_path
-        bits, group_size, sym = 4, 128, True
-        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
-        tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
-        layer_config = {
-            r"model\.decoder\.layers\.(?:[0-9]|1[0-1])\.self_attn\.q_proj": {"bits": 16},
-            ##"model.decoder.layers.1.self_attn.k_proj": {"bits": 16}
-        }
-        autoround = AutoRound(
-            model,
-            tokenizer=tokenizer,
-            bits=bits,
-            group_size=group_size,
-            sym=sym,
-            iters=2,
-            seqlen=2,
-            dataset=dataloader,
-            layer_config=layer_config,
-        )
-        autoround.quantize()
-        quantized_model_path = self.save_folder
-
-        _, quantized_model_path = autoround.save_quantized(
-            output_dir=quantized_model_path, format="auto_round", inplace=True, return_folders=True
-        )
-        quantization_config = AutoRoundConfig()
-
-        model = AutoModelForCausalLM.from_pretrained(
-            quantized_model_path, device_map="auto", quantization_config=quantization_config
-        )
-        tokenizer = AutoTokenizer.from_pretrained(quantized_model_path)
-        text = "There is a girl who likes adventure,"
-        inputs = tokenizer(text, return_tensors="pt").to(model.device)
-        res = tokenizer.decode(model.generate(**inputs, max_new_tokens=5)[0])
-        print(res)
-
-    def test_fallback_layers_regex_round(self, tiny_opt_model_path, dataloader):
-        model_name = tiny_opt_model_path
-        bits, group_size, sym = 4, 128, True
-        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
-        tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
-        layer_config = {
-            r"model\.decoder\.layers\.(?:[0-9]|1[0-1])\.self_attn\.q_proj": {"bits": 16},
-            r"model.decoder.layers.1.self_attn.k_proj": {"bits": 16},
-        }
-        autoround = AutoRound(
-            model,
-            tokenizer=tokenizer,
-            bits=bits,
-            group_size=group_size,
-            sym=sym,
-            iters=2,
-            seqlen=2,
-            dataset=dataloader,
-            layer_config=layer_config,
-        )
-        autoround.quantize()
-        quantized_model_path = self.save_folder
-
-        _, quantized_model_path = autoround.save_quantized(
-            output_dir=quantized_model_path, format="auto_round", inplace=True, return_folders=True
-        )
-        quantization_config = AutoRoundConfig()
-
-        model = AutoModelForCausalLM.from_pretrained(
-            quantized_model_path, device_map="auto", quantization_config=quantization_config
-        )
-        tokenizer = AutoTokenizer.from_pretrained(quantized_model_path)
-        text = "There is a girl who likes adventure,"
-        inputs = tokenizer(text, return_tensors="pt").to(model.device)
-        res = tokenizer.decode(model.generate(**inputs, max_new_tokens=5)[0])
-        print(res)
-
-    def test_dequant_fp8_weight(self):
-        from auto_round.utils.model import _dequant_fp8_linear_weight
-
-        # test high precision weight (should skip LUT and not crash)
-        weight = torch.randn(587, 7168)
-        weight_scale = torch.randn(5, 56)
-        block_size = [128, 128]
-        dequant_weight = _dequant_fp8_linear_weight(weight, weight_scale, block_size)
-        assert dequant_weight.shape.numel() == 4207616
-
-        # test low precision weight (should use LUT)
-        weight = torch.randint(0, 255, (587, 7168), dtype=torch.uint8)
-        weight_scale = torch.randn(5, 56)
-        block_size = [128, 128]
-        dequant_weight = _dequant_fp8_linear_weight(weight, weight_scale, block_size)
-        assert dequant_weight.shape.numel() == 4207616
-
-        # test experts are stacked.
-        weight = torch.randint(0, 255, [32, 5760, 1440], dtype=torch.uint8)
-        weight_scale = torch.randn([32, 5760, 90])
-        block_size = [1, 16]
-        dequant_weight = _dequant_fp8_linear_weight(weight, weight_scale, block_size)
-        assert len(dequant_weight.shape) == 3
-        assert dequant_weight.shape[0] == 32
-        assert dequant_weight.shape.numel() == 32 * 5760 * 1440
-
-    def test_dequant_g2_fp8_weight(self):
-        """Test Gaudi2-specific dequantization redirection logic in convert_module_to_hp_if_necessary."""
-        from unittest.mock import MagicMock, patch
-
-        from auto_round.utils.model import (
-            _dequant_fp8_linear_weight,
-            check_and_mark_quantized_module,
-            convert_module_to_hp_if_necessary,
-        )
-
-        # Test 1: Verify Standard e4m3fn native dequantization result
-        weight_uint8 = torch.tensor([[126, 0]], dtype=torch.uint8)
-        weight_scale = torch.ones(1, 1, dtype=torch.bfloat16)
-        block_size = [1, 2]
-
-        dq_weight = _dequant_fp8_linear_weight(weight_uint8, weight_scale, block_size)
-        assert dq_weight.dtype == torch.bfloat16
-        assert abs(dq_weight[0, 0].item() - 448.0) < 1e-3
-
-        # Test 2: Verify Gaudi2 redirection in convert_module_to_hp_if_necessary
-        mock_layer = MagicMock()
-        mock_layer.in_features = 2
-        mock_layer.out_features = 1
-        mock_layer.bias = None
-        mock_layer.weight = weight_uint8
-        mock_layer.weight_scale = weight_scale
-        mock_layer.block_size = block_size
-        mock_layer.data_type = "fp8_e4m3fn"
-        mock_layer.__class__.__name__ = "FP8Linear"
-        # Explicitly set compressor to None to prevent MagicMock auto-creating it
-        mock_layer.compressor = None
-
-        # mock_layer.to should return a mock layer that we can check
-        mock_layer.to.return_value = mock_layer
-
-        with patch("auto_round.utils.device.is_gaudi2", return_value=True):
-            check_and_mark_quantized_module(mock_layer)
-            convert_module_to_hp_if_necessary(mock_layer, device="hpu")
-            # Verify it was moved to meta
-            mock_layer.to.assert_called_with("meta")
-
-        with patch("auto_round.utils.device.is_gaudi2", return_value=False):
-            # Reset mock
-            mock_layer.to.reset_mock()
-            check_and_mark_quantized_module(mock_layer)
-            convert_module_to_hp_if_necessary(mock_layer, device="hpu")
-            # Verify it was moved to meta
-            mock_layer.to.assert_called_with("meta")
-
-    def test_mixed_bit_setting(self, tiny_opt_model_path):
-        model_name = tiny_opt_model_path
-        layer_config = {"model.decoder.layers.1.fc1": {"bits": 8, "act_bits": 8}}
-        ar = AutoRound(model_name, data_type="mx_fp4", act_bits=4, iters=0, layer_config=layer_config)
-        ar.quantize()
-        layer_config = ar.layer_config
-        if (
-            layer_config["model.decoder.layers.1.fc1"]["bits"] != 8
-            or layer_config["model.decoder.layers.1.fc1"]["act_bits"] != 8
-        ):
-            raise ValueError("mixed bits is not correct")
-
-    def test_quant_lm_head(self, tiny_untied_qwen_model_path):
-        model_name = tiny_untied_qwen_model_path
-        ar = AutoRound(
-            model_name,
-            quant_lm_head=True,
-            iters=0,
-            seqlen=8,
-            nsamples=1,
-            disable_opt_rtn=True,
-        )
-        _, quantized_model_path = ar.quantize_and_save(output_dir=self.save_folder, format="auto_round")
-        model = AutoModelForCausalLM.from_pretrained(quantized_model_path, device_map="cpu")
-        assert "lm_head" in model.config.quantization_config.extra_config
-        assert model.config.quantization_config.extra_config["lm_head"]["bits"] == 4
-
-    def test_quant_lm_head_layer_config(self, tiny_untied_qwen_model_path):
-        model_name = tiny_untied_qwen_model_path
-        layer_config = {"lm_head": {"bits": 4}}
-        ar = AutoRound(
-            model_name,
-            quant_lm_head=True,
-            iters=0,
-            seqlen=8,
-            nsamples=1,
-            disable_opt_rtn=True,
-            layer_config=layer_config,
-        )
-        _, quantized_model_path = ar.quantize_and_save(output_dir=self.save_folder, format="auto_round")
-        model = AutoModelForCausalLM.from_pretrained(quantized_model_path, device_map="cpu")
-        assert "lm_head" in model.config.quantization_config.extra_config
-        assert model.config.quantization_config.extra_config["lm_head"]["bits"] == 4
-
-    def test_compressor(self, tiny_qwen_vl_model_path):
-        model_name = tiny_qwen_vl_model_path
-        ar = AutoRound(model_name, enable_adam=True)
-        assert ar.optimizer == torch.optim.AdamW
-        assert ar.mllm
-
-    def test_attention_mask_in_dataset(self):
-        from transformers import AutoTokenizer
-
-        model_name = qwen_name_or_path
-        tokenizer = AutoTokenizer.from_pretrained(model_name)
-        text = ["haha", "hello world"]
-        res = tokenizer(text, return_tensors="pt", max_length=8, padding="max_length", truncation=True)
-        data = [res.data]
-
-        text = ["qudd", "hfd"]
-        res = tokenizer(text, return_tensors="pt", max_length=8, padding="max_length", truncation=True)
-        data.append(res.data)
-        from auto_round import AutoRound
-
-        ar = AutoRound(model_name, iters=1, dataset=data, seqlen=8)
-        ar.quantize()
-
-    def test_attention_mask_via_tokenize_in_dataset(self):
-        from transformers import AutoTokenizer
-
-        model_name = qwen_name_or_path
-        tokenizer = AutoTokenizer.from_pretrained(model_name)
-        text = ["haha", "hello world"]
-        res = tokenizer(text, return_tensors="pt", max_length=8, padding="max_length", truncation=True)
-        res.data.pop("attention_mask")
-        data = [res.data]
-
-        text = ["qudd", "hfd"]
-        res = tokenizer(text, return_tensors="pt", max_length=8, padding="max_length", truncation=True)
-        res.data.pop("attention_mask")
-        data.append(res.data)
-        from auto_round import AutoRound
-
-        ar = AutoRound(model_name, iters=1, dataset=data, seqlen=8)
-        ar.quantize()
-
-    def test_low_cpu_mem_usage(self, tiny_opt_model_path, dataloader):
-        bits, group_size = 4, 32
-        model_name = tiny_opt_model_path
-        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
-        tokenizer = AutoTokenizer.from_pretrained(model_name, trust_remote_code=True)
-        quantized_model_path = self.save_folder
-        autoround = AutoRound(
-            model,
-            tokenizer,
-            bits=bits,
-            group_size=group_size,
-            iters=2,
-            seqlen=10,
-            dataset=dataloader,
-            low_cpu_mem_usage=True,
-            device_map="cpu",
-        )
-        autoround.quantize_and_save(output_dir=quantized_model_path, format="auto_round")
-
-    def test_create_adam(self):
-        model_name = qwen_name_or_path
-        from auto_round import AutoRound
-
-        ar = AutoRound(model=model_name, enable_adam=True)
