+++ test/test_cpu/export/test_export.py
@@ -290,373 +290,3 @@ class TestAutoRound:
         assert "model.decoder.layers.8.self_attn.k_proj.weight_scale" in f.keys()
         assert f.get_tensor("model.decoder.layers.5.self_attn.v_proj.input_scale").shape == torch.Size([1])
         assert f.get_tensor("model.decoder.layers.5.self_attn.v_proj.weight").dtype == torch.float8_e4m3fn
-
-    def test_static_fp8_attn(self):
-        import os
-
-        from safetensors import safe_open
-
-        model_name = self.model_name
-        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
-        autoround = AutoRound(
-            model,
-            self.tokenizer,
-            iters=0,
-            nsamples=2,
-            seqlen=2,
-            scheme="FP8_STATIC",
-            static_attention_dtype="fp8",
-        )
-        quantized_model_path = self.save_dir
-        _, quantized_model_path = autoround.quantize_and_save(output_dir=quantized_model_path, format="auto_round")
-        f = safe_open(os.path.join(quantized_model_path, "model.safetensors"), framework="pt")
-        assert "model.decoder.layers.8.self_attn.k_proj.input_scale" in f.keys()
-        assert "model.decoder.layers.8.self_attn.k_proj.weight_scale" in f.keys()
-        assert f.get_tensor("model.decoder.layers.5.self_attn.v_proj.input_scale").shape == torch.Size([1])
-        assert f.get_tensor("model.decoder.layers.5.self_attn.v_proj.weight").dtype == torch.float8_e4m3fn
-        check_attrs = ["k_scale", "v_scale", "q_scale"]
-        for attr in check_attrs:
-            weight_name = f"model.decoder.layers.8.self_attn.{attr}"
-            assert weight_name in f.keys()
-            assert f.get_tensor(weight_name).shape == torch.Size([1])
-            assert f.get_tensor(weight_name).dtype == torch.float32 or f.get_tensor(weight_name).dtype == torch.bfloat16
-
-    def test_awq_lmhead_export(self, dataloader):
-        bits, sym, group_size = 4, False, 128
-        model_name = get_model_path("microsoft/phi-4")
-        layer_config = {
-            "lm_head": {"bits": 4},  # set lm_head quant
-            "layer": {"bits": 16},
-        }
-        autoround = AutoRound(
-            model=model_name,
-            bits=bits,
-            group_size=group_size,
-            sym=sym,
-            iters=2,
-            nsamples=2,
-            seqlen=2,
-            layer_config=layer_config,
-            dataset=dataloader,
-        )
-        quantized_model_path = self.save_dir
-        compressed_model, _ = autoround.quantize_and_save(output_dir=quantized_model_path, format="auto_awq")
-        lm_head = compressed_model.lm_head
-        from auto_round.export.export_to_awq.utils import WQLinear_GEMM
-
-        assert isinstance(lm_head, WQLinear_GEMM), "Illegal AWQ quantization for lm_head layer"
-
-    def test_gptq_lmhead_export(self, dataloader):
-        bits, sym, group_size = 4, True, 128
-        # Note that, to save UT tuning time, the local model is intentionally kept lightweight, using only 2 hidden layers.
-        model_name = get_model_path("microsoft/phi-4")
-        layer_config = {
-            "lm_head": {"bits": 4},  # set lm_head quant
-            "layer": {"bits": 16},
-        }
-        autoround = AutoRound(
-            model=model_name,
-            bits=bits,
-            group_size=group_size,
-            sym=sym,
-            nsamples=2,
-            iters=2,
-            seqlen=2,
-            layer_config=layer_config,
-            dataset=dataloader,
-        )
-        quantized_model_path = self.save_dir
-        compressed_model, quantized_model_path = autoround.quantize_and_save(
-            output_dir=quantized_model_path, format="auto_gptq"
-        )
-        lm_head = compressed_model.lm_head
-        assert hasattr(lm_head, "bits") and lm_head.bits == 4, "Illegal GPTQ quantization for lm_head layer"
-        quantization_config = AutoRoundConfig()
-        model = AutoModelForCausalLM.from_pretrained(
-            quantized_model_path, device_map="cpu", quantization_config=quantization_config
-        )
-        tokenizer = AutoTokenizer.from_pretrained(quantized_model_path)
-        text = "There is a girl who likes adventure,"
-        inputs = tokenizer(text, return_tensors="pt").to(model.device)
-        res = tokenizer.decode(model.generate(**inputs, max_new_tokens=5)[0])
-        print(res)
-
-    def test_export_format(self):
-        from auto_round.formats import get_formats
-
-        autoround = AutoRound(
-            self.model_name,
-            scheme="FP8_STATIC",
-        )
-        autoround.post_init()
-        format_list = get_formats("auto_round, llm_compressor, auto_round:llm_compressor", autoround)
-        assert len(format_list) == 3
-        assert format_list[0].output_format == "auto_round"
-        assert format_list[0].get_backend_name() == "auto_round:fp8_static"
-        assert format_list[1].output_format == "llm_compressor"
-        assert format_list[1].get_backend_name() == "llm_compressor:fp8_static"
-        assert format_list[2].output_format == "auto_round"
-        assert format_list[2].get_backend_name() == "auto_round:llm_compressor:fp8_static"
-
-        autoround = AutoRound(
-            self.model_name,
-            scheme="W4A16",
-        )
-        autoround.post_init()
-        format_list = get_formats("auto_round:auto_awq, auto_gptq", autoround)
-        assert format_list[0].output_format == "auto_round"
-        assert format_list[0].get_backend_name() == "auto_round:auto_awq"
-        assert format_list[1].output_format == "auto_gptq"
-        assert format_list[1].get_backend_name() == "auto_gptq"
-
-        autoround = AutoRound(
-            model=self.model_name,
-            scheme="INT8",
-        )
-        autoround.post_init()
-        format_list = get_formats("llm_compressor, auto_round:llm_compressor", autoround)
-        assert format_list[0].output_format == "llm_compressor"
-        assert format_list[0].get_backend_name() == "llm_compressor:int8_w8a8"
-        assert format_list[1].output_format == "auto_round"
-        assert format_list[1].get_backend_name() == "auto_round:llm_compressor:int8_w8a8"
-
-        # Verify backward compatibility: INT8_W8A8 (old name) produces identical formats to INT8
-        autoround_old = AutoRound(
-            model=self.model_name,
-            scheme="INT8_W8A8",
-        )
-        autoround_old.post_init()
-        format_list_old = get_formats("llm_compressor, auto_round:llm_compressor", autoround_old)
-        assert format_list_old[0].output_format == "llm_compressor"
-        assert format_list_old[0].get_backend_name() == "llm_compressor:int8_w8a8"
-        assert format_list_old[1].output_format == "auto_round"
-        assert format_list_old[1].get_backend_name() == "auto_round:llm_compressor:int8_w8a8"
-
-    def test_export_format_with_scheme(self, tiny_qwen_model_path):
-        from auto_round.formats import get_formats
-
-        ar = AutoRound(
-            model=tiny_qwen_model_path,
-            scheme="W4A16",
-            bits=2,
-            group_size=32,
-            sym=True,
-        )
-        ar.post_init()
-        with pytest.raises(ValueError, match="auto_awq format support quantization scheme with W4A16 but got bits=2"):
-            get_formats("auto_round:auto_awq", ar)
-
-        with pytest.raises(ValueError, match="but got bits=2, data_type=int"):
-            get_formats("auto_round:llm_compressor", ar)
-
-        ar = AutoRound(
-            model=tiny_qwen_model_path,
-            scheme="FP8_STATIC",
-            bits=4,
-            group_size=32,
-            sym=True,
-        )
-        ar.post_init()
-        with pytest.raises(ValueError, match="but got data_type=fp, bits=4"):
-            get_formats("auto_round:llm_compressor", ar)
-
-        ar = AutoRound(
-            model=tiny_qwen_model_path,
-            scheme="w2a16",
-            bits=4,
-            group_size=256,
-            sym=True,
-        )
-        ar.post_init()
-        get_formats("auto_round:auto_awq", ar)
-
-    def test_autoawq_qwen3_vl_infer(self, dataloader):
-        model_path = get_model_path("Qwen/Qwen3-VL-2B-Instruct")
-        autoround = AutoRound(
-            model=model_path,
-            scheme="W4A16",
-            iters=0,
-            seqlen=2,
-            batch_size=1,
-            dataset=dataloader,
-        )
-        quantized_model_path = self.save_dir
-        _, quantized_model_path = autoround.quantize_and_save(
-            output_dir=quantized_model_path, inplace=False, format="auto_awq"
-        )
-
-        # Check items of modules_to_not_convert in quantization config
-        quantization_config_path = f"{quantized_model_path}/quantization_config.json"
-        with open(quantization_config_path, "r") as f:
-            quantization_config = json.load(f)
-        modules_to_not_convert = quantization_config.get("modules_to_not_convert", [])
-        assert (
-            "model.visual.merger.linear_fc2" in modules_to_not_convert
-        ), f"'model.visual.merger.linear_fc2' should be in modules_to_not_convert. Got: {modules_to_not_convert}"
-        assert (
-            "model.visual.merger.linear_fc1" in modules_to_not_convert
-        ), f"'model.visual.merger.linear_fc1' should be in modules_to_not_convert. Got: {modules_to_not_convert}"
-        assert (
-            "model.visual.blocks" in modules_to_not_convert
-        ), f"'model.visual.blocks' should be in modules_to_not_convert. Got: {modules_to_not_convert}"
-
-    @pytest.mark.parametrize(
-        "iters,use_dataloader,scheme",
-        [
-            (0, False, "INT8"),  # RTN with new scheme name
-            (1, True, "INT8"),  # tuning with new scheme name
-            (0, False, "INT8_W8A8"),  # RTN with old scheme name (backward compat)
-        ],
-        ids=["rtn", "tuning", "rtn-old-scheme"],
-    )
-    def test_llmc_dynamic_wint8aint8_export(self, iters, use_dataloader, scheme, dataloader):
-        from safetensors import safe_open
-
-        dataset = dataloader if use_dataloader else None
-        autoround = AutoRound(
-            self.model_name,
-            iters=iters,
-            nsamples=2,
-            seqlen=2,
-            dataset=dataset,
-            scheme=scheme,
-        )
-        quantized_model_path = self.save_dir
-        _, quantized_model_path = autoround.quantize_and_save(output_dir=quantized_model_path, format="llm_compressor")
-        with safe_open(os.path.join(quantized_model_path, "model.safetensors"), framework="pt") as f:
-            assert "model.decoder.layers.8.self_attn.k_proj.weight_scale" in f.keys()
-            assert f.get_tensor("model.decoder.layers.5.self_attn.v_proj.weight").dtype == torch.int8
-        shutil.rmtree(quantized_model_path, ignore_errors=True)
-
-    @pytest.mark.parametrize(
-        "scheme,bits,group_size,sym",
-        [
-            ("W4A16", 4, 128, True),
-            ("W4A16", 4, -1, True),
-            ("W8A16", 8, -1, True),
-        ],
-    )
-    def test_llmc_wint_a16_export(self, scheme, bits, group_size, sym):
-        from safetensors import safe_open
-
-        autoround = AutoRound(
-            self.model_name,
-            iters=2,
-            nsamples=2,
-            seqlen=2,
-            scheme=scheme,
-            bits=bits,
-            group_size=group_size,
-            sym=sym,
-        )
-        quantized_model_path = self.save_dir
-        _, quantized_model_path = autoround.quantize_and_save(output_dir=quantized_model_path, format="llm_compressor")
-        with safe_open(os.path.join(quantized_model_path, "model.safetensors"), framework="pt") as f:
-            # weights must be packed as int32 (compressed-tensors stores both int4 and int8 as torch.int32)
-            weight = f.get_tensor("model.decoder.layers.5.self_attn.v_proj.weight_packed")
-            assert weight.dtype == torch.int32, f"Expected int32 weight for {scheme}, got {weight.dtype}"
-            # weight_scale must be present and be a float tensor
-            scale_key = "model.decoder.layers.8.self_attn.k_proj.weight_scale"
-            assert scale_key in f.keys(), f"Missing {scale_key} for {scheme} export"
-            scale = f.get_tensor(scale_key)
-            assert scale.dtype in (
-                torch.float32,
-                torch.float16,
-                torch.bfloat16,
-            ), f"Expected float weight_scale for {scheme}, got {scale.dtype}"
-            # No input_scale should be present for weight-only quantization
-            input_scale_keys = [k for k in f.keys() if k.endswith(".input_scale")]
-            assert (
-                len(input_scale_keys) == 0
-            ), f"Expected no input_scale for weight-only {scheme}, but found: {input_scale_keys[:5]}"
-        shutil.rmtree(quantized_model_path, ignore_errors=True)
-
-
-@pytest.mark.parametrize(
-    "format_name,export_module,sym",
-    [
-        ("auto_gptq", autogptq_export, False),
-        ("auto_awq", awq_export, False),
-        ("auto_round", autoround_export, True),
-    ],
-)
-def test_weight_only_exports_pack_serially(tiny_opt_model_path, tmp_path, monkeypatch, format_name, export_module, sym):
-    autoround = AutoRound(
-        tiny_opt_model_path,
-        bits=4,
-        group_size=128,
-        sym=sym,
-        iters=0,
-        disable_opt_rtn=True,
-    )
-    autoround.quantize()
-    forbid_threaded_packing(monkeypatch, export_module)
-    autoround.save_quantized(output_dir=tmp_path, inplace=False, format=format_name)
-    assert os.path.exists(os.path.join(tmp_path, "config.json"))
-
-
-def test_fp8_autoround_export_packs_serially(tiny_opt_model_path, tmp_path, monkeypatch):
-    from safetensors import safe_open
-
-    autoround = AutoRound(
-        tiny_opt_model_path,
-        bits=8,
-        group_size=-1,
-        iters=0,
-        scheme="FP8_STATIC",
-        nsamples=2,
-        seqlen=2,
-        static_kv_dtype="fp8",
-    )
-    autoround.quantize()
-    forbid_threaded_packing(monkeypatch, autoround_fp8_export)
-    autoround.save_quantized(output_dir=tmp_path, format="auto_round")
-    with safe_open(os.path.join(tmp_path, "model.safetensors"), framework="pt") as f:
-        assert "model.decoder.layers.0.self_attn.k_proj.weight_scale" in f.keys()
-
-
-@pytest.mark.parametrize("low_cpu_mem_usage", [True, False])
-def test_immediate_saving_mode(tiny_opt_model_path, tmp_path, low_cpu_mem_usage, caplog):
-    """Verify that immediate_saving (triggered by low_cpu_mem_usage) produces a complete model output."""
-    import logging
-
-    output_dir = str(tmp_path / "output")
-    with caplog.at_level(logging.DEBUG, logger="auto_round"):
-        autoround = AutoRound(
-            tiny_opt_model_path,
-            scheme="MXFP4",
-            iters=2,
-            seqlen=2,
-            nsamples=2,
-            low_cpu_mem_usage=low_cpu_mem_usage,
-        )
-        _, quantized_model_path = autoround.quantize_and_save(output_dir=output_dir, format="llm_compressor")
-
-    # No spurious "already exists" warning should be emitted
-    conflict_messages = [r.message for r in caplog.records if "already exists" in r.message]
-    assert len(conflict_messages) == 0, f"Unexpected conflict warnings: {conflict_messages}"
-
-    # All essential files must exist regardless of immediate_saving mode
-    assert os.path.exists(os.path.join(quantized_model_path, "config.json")), "config.json missing"
-    assert os.path.exists(
-        os.path.join(quantized_model_path, "quantization_config.json")
-    ), "quantization_config.json missing"
-
-    # Exactly 1 safetensors shard for this tiny model
-    safetensor_files = [f for f in os.listdir(quantized_model_path) if f.endswith(".safetensors")]
-    assert len(safetensor_files) == 1, f"Expected 1 safetensors file, got {len(safetensor_files)}: {safetensor_files}"
-
-    # Tokenizer files must be present
-    assert os.path.exists(os.path.join(quantized_model_path, "tokenizer_config.json")), "tokenizer_config.json missing"
-
-    # Total file count: config.json, quantization_config.json, model.safetensors,
-    # generation_config.json, tokenizer.json, tokenizer_config.json = 6
-    all_files = os.listdir(quantized_model_path)
-    assert len(all_files) == 6, f"Expected 6 files, got {len(all_files)}: {sorted(all_files)}"
-
-    # Verify weights are loadable and non-empty
-    from safetensors import safe_open
-
-    with safe_open(os.path.join(quantized_model_path, safetensor_files[0]), framework="pt") as f:
-        keys = f.keys()
-        assert len(keys) > 0, "Safetensors file has no tensors"
